添加了markdown功能及完善blogdetail

+++ webpack/plugins/change-antd-theme.js
@@ -1,16 +1,20 @@
 var fs = require('fs');
 var path = require('path');
 function ChangeAntdPlugin(options) {
-  this.context = options.context
+  this.context = options.context,
+  this.todo = true
 }
 
 ChangeAntdPlugin.prototype.apply = function(compiler) {
   var that = this;
-  compiler.plugin('optimize', function(compilation, callback) {
-    var antdLess = path.join(that.context, './app/constants/', 'antd.less');
-    var defaultLess = path.join(that.context, './node_modules/antd/lib/style/themes', 'default.less');
-    fs.unlinkSync(defaultLess);
-    fs.linkSync(antdLess,defaultLess);
+  compiler.plugin('compile', function(compilation, callback) {
+    if (that.todo) {
+      var antdLess = path.join(that.context, './app/constants/', 'antd.less');
+      var defaultLess = path.join(that.context, './node_modules/antd/lib/style/themes', 'default.less');
+      fs.unlinkSync(defaultLess);
+      fs.linkSync(antdLess,defaultLess);
+      that.todo = false
+    }
   })
 }
